hackathon/smart_travel_journey_planner/requirements.txt
google-adk>=1.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tzdata
jinja2>=3.1.3
python-multipart>=0.0.9
sse-starlette>=2.0.0
tenacity>=8.2.0
httpx
google-auth>=2.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
itsdangerous
workspace-mcp